EDMT-302 refactor : mock 핸들러 및 초기 msw 설정 코드 shared 하위 폴더로 이동
