+++ src/shared/mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw/browser';
 
-import { handlers } from './handlers/index';
+import { handlers } from './handlers';
 
 export const worker = setupWorker(...handlers);
